Crud upgrades

+++ back-end/src/main/java/com/bimi/bankingsystem/repository/LoanRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LoanRepository extends JpaRepository<Loan,Integer> {
+public interface LoanRepository extends JpaRepository<Loan,Long> {
 }
